Turn off hard-coded UTF-8 parsing of caught errors
Test turning off the encoding for caught errors to see if this allows for better handling of errors related to source data that might not be in the UTF-8 charset. Per the {httr} documentation for `content()`, this change should switch to dynamically determining the encoding from the HTTP response header instead of assuming it's UTF-8.